+++ assets/javascripts/kitten/components/cards/crowdfunding-card.js
@@ -269,7 +269,7 @@ class CrowdfundingCardComponent extends Component {
   }
 
   renderProgress() {
-    if (!this.props.progress) return
+    if (this.props.progress === false) return
 
     const progressValue = this.props.loading
       ? 65
@@ -427,6 +427,7 @@ const styles = {
     },
   },
   stroke: {
+    flexShrink: 0,
     margin: `5px ${COMPONENT_GUTTER}px 5px 0`,
 
     loading: {
@@ -517,7 +518,7 @@ CrowdfundingCardComponent.defaultProps = {
   info1: null,
   info2: null,
   info3: null,
-  progress: 0,
+  progress: false,
   state: null,
   loading: false,
 }