+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `FilterIcon` component.
 - Feature: Update all buttons colors.
 
 ## [20.0.2] - 2018-03-05