+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `CrowdfundingCard` when the progress value is zero.
 - Feature: Add new `FilterIcon` component.
 - Feature: Update all buttons colors.