+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Fix `CrowdfundingCard` when the progress value is zero.
+- Feature: Add progress percentage in `CrowdfundingCard`.
+- Feature: Add `azote` modifier on `Button` and `ButtonIcon` components.
+- Fix: Update primary colors.
 - Feature: Add new `FilterIcon` component.
 - Feature: Update all buttons colors.