+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Fix `TagButton` colors.
+- Fix: Update `k-Checkbox__label` with flex.
+- Feature: Add `ScrollableContainer`.
+- Fix: Fix `CrowdfundingCard` when the progress value is zero.
+- Feature: Add progress percentage in `CrowdfundingCard`.
 - Feature: Add `azote` modifier on `Button` and `ButtonIcon` components.
 - Fix: Update primary colors.
 - Feature: Add new `FilterIcon` component.